pyproject.toml
[tool.poetry]
name = "InteractiveTeacher"
version = "0.1.0"
description = ""
authors = ["Fernando Gonzalez <fernando.gonzalez@kavak.com>","Marco Andrade <marco.andrade@kavak.com>","Milan Sahu<milan.sahu@kavak.com>","Raul Perez (Rulo) <raul.perez@kavak.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
python-dotenv = "^1.0.0"
requests = "^2.31.0"
loguru = "^0.7.2"
sqlalchemy = "^2.0.23"
pydantic = "^2.4.2"
psycopg2-binary = "^2.9.9"
langchain = "^0.1.16"
boto3 = "^1.34.11"
openai = "^0.28.1"
langchain-core = "^0.1.46"
twilio = "^9.0.5"
h11 = "^0.14.0"
pydantic-core = "^2.18.1"
annotated-types = "^0.6.0"
langchain-google-vertexai = "^1.0.1"
langchainhub = "^0.1.15"
pillow = "^10.3.0"
langchain-google-genai = "^1.0.3"
wikipedia = "^1.4.0"
pdf2image = "^1.17.0"
google-cloud-texttospeech = "^2.16.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
pytest = "^7.4.2"
coverage = "^7.3.1"
pytest-asyncio = "^0.21.1"
dbmate = "^0.0.1"
pytest-watch = "^4.2.0"
black = "^24.1.1"
locust = "^2.21.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "pipy"
url = "https://pypi.python.org/simple"
priority = "primary"
